Refactor Sidebar Conditionals
Changed our test for an empty sidebar. We used to pull the sidebar into a buffer and test it to see if it's an empty string, but that was causing trouble with certain sidebar widgets.

I switched things back to a more Wordpress-y way, using the is_active_sidebar() function to see if a sidebar was active on the page, and had any widgets to render.