Add Mobile Menu Items
The README file for the ASU Header says you can insert your site's menu into the ASU mobile menu by defining it in a JSON object.  For the moment, I am using a static object, to make sure this all works. We could gather the menu items programmatically in the future.