Hide Corona Virus Banner by Default
Updated the code in our page layouts that displays the global ASU Coronavirus banner so that "display: none" is the default and the banner is hidden by default.

This way, we can push our other updates to production, and then just go to the SOS website and use the customizer to override "display: none" to "display: block" for that class in order to turn the banner back on for the school site.