Remove WordPress Default oEmbed <link> Tags
It appears that our LinkedIn previews have not been correct because LinkedIn will ignore the "og:title" we have set if there is an oEmbed tag to use. As it turns out, WordPress will, by default, add oEmbed tags to the <head> of every page. This code removes that action from WordPress, resulting in pages that do NOT have oEmbed tags and should preview correctly when shared via LinkedIn.